refactor: compose screen
